+++ compiler/src/iree/compiler/Codegen/Common/TransformExtensions/CommonExtensions.cpp
@@ -967,6 +967,7 @@ transform_dialect::TestGpuVectorDistribution::applyToOne(
   populateGPUDistributionLayoutAttrPatterns(laneId, patterns);
   populateGPUReductionDistributionPatterns(patterns);
   populateGPUDistributeNestedLayoutAttrPatterns(laneId, patterns);
+  populateGPUDistributeNestedLayoutContractAMDGPUPatterns(laneId, patterns);
   if (getExperimental())
     populateGPULayoutResolutionDistributionPatterns(patterns);
   distributeVectorOps(target, patterns, options);
